Explicit Interface

Explicit interface

in addition to the standard logging interface, it is possible to log data to TensorBoard using the functions documented below. All the functions accept take as first argument a TBLogger object and as the second argument a String as the tag under which the data will be logged.

Scalar backend

TensorBoardLogger.log_value — Function.

log_value(logger, name, value; step=step(logger))

Logs a Floating-point variable with name name at step step. If value is a complex number, the real and imaginary part are logged separately.

source

Distributions backend

TensorBoardLogger.log_histogram — Function.

log_histogram(logger, name, (bins,weights); step=step(logger))

Logs a histogram under the tag name on given step. The histogram must be passed as a tuple holding the N+1 bin edges and the height of the N bins.

You can also pass the raw data, and a binning algorithm from StatsBase.jl will be used to bin the data.

source

log_histogram(logger, name, data::Vector; step=step(logger))

Bins the values found in data and logs them as an histogram under the tag name.

source

TensorBoardLogger.log_vector — Function.

log_vector(logger, name, data::Vector; step=step(logger))

Logs the vector found in data as an histogram under the name name.

source

# Text backend

TensorBoardLogger.log_text — Function.

log_text(logger::TBLogger, name::String, text::Any; step=step(logger))

Logs text with name name at step step

text: If text is a 2-D or 3-D Array, it will be rendered as a table or a list. Any other data will be represented as string

source

Images backend

TensorBoardLogger.log_image — Function.

log_image(logger::TBLogger, name::AbstractString, imgArray::AbstractArray, format::ImageFormat; step=step(logger))

Log an image using image data and format

imgArray: image data. A 1-D, 2-D or 3-D Array of pixel values. pixel values can be Real [0, 1] or Integer[0, 255]

format: format of the image. It can be one of {L, LN, NL, CL, LC, NCL, NLC, CLN, LCN, HW, WH, HWC, WHC, CHW, CWH, HWN, WHN, NHW, NWH, HWCN, WHCN, CHWN, CWHN, NHWC, NWHC, NCHW, NCWH}

L: Length

C: Channel/Color

H: Height

W: Width

N: Observation

source

TensorBoardLogger.log_images — Function.

log_images(logger::TBLogger, name::AbstractString, imgArrays::AbstractArray, format::ImageFormat; step=step(logger))

Log multiple images using Array of images and format

imgArrays: Array of images, e.g. Array{Array{Float64, 3}, 1}. Array of images can be multidimensional.

format: format which applies to each image in the Array of images. It can be one of {L, LN, NL, CL, LC, NCL, NLC, CLN, LCN, HW, WH, HWC, WHC, CHW, CWH, HWN, WHN, NHW, NWH, HWCN, WHCN, CHWN, CWHN, NHWC, NWHC, NCHW, NCWH}

source

Audio backend

TensorBoardLogger.log_audio — Function.

log_audio(logger::TBLogger, name::AbstractString, samples::AbstractArray, samplerate::Real; step=step(logger))

Logs an audio clip with name name at step step

samples: Array of samples N*C where N = number of samples and C = number of channels

samplerate: Sampling rate or Sampling frequency: a Real value

source

TensorBoardLogger.log_audios — Function.

log_audios(logger::TBLogger, name::AbstractString, samples::AbstractArray, samplerate::Real; step=step(logger))

Logs multiple audio clips at step step

samples: Array of audio clips which are Arrays of samples N*C where N = number of samples and C = number of channel

samplerate: Sampling rate or Sampling frequency: a Real value same for all clips

source