Hyperparameter logging

In additition to logging the experiments, you may wish to also visualise the effect of hyperparameters on some plotted metrics. This can be done by logging the hyperparameters via the write_hparams! function, which takes a dictionary mapping hyperparameter names to their values (currently limited to Real, Bool or String types), along with the names of any metrics that you want to view the effects of.

You can see how the HParams dashboard in Tensorboard can be used to tune hyperparameters on the tensorboard website.

API

TensorBoardLogger.write_hparams! — Function

write_hparams!(logger::TBLogger, hparams::Dict{String, Any}, metrics::AbstractArray{String})

Writes the supplied hyperparameters to the logger, along with noting all metrics that should be tracked.

The value of hparams can be a String, Bool or a subtype of Real. All Real values are converted to Float64 when writing the logs.

metrics should be a list of tags, which correspond to scalars that have been logged. Tensorboard will automatically extract the latest metric logged to use for this value.